Update field names and correcting assumptions/usage

+++ ADRIA/src/io/result_io.jl
@@ -212,7 +212,7 @@ function setup_logs(
         UInt16,
         rank_dims...;
         name="rankings",
-        fill_value=0,
+        fill_value=UInt16(0),
         fill_as_missing=false,
         path=log_fn,
         chunks=(rank_dims[1:3]..., batch_size),
